@@ -51,54 +57,62 @@
   "dependencies": {
     "@ledgerhq/hw-transport": "4.24.0",
     "@ledgerhq/hw-transport-node-hid": "4.24.0",
-    "aws-sdk": "^2.395.0",
+    "aws-sdk": "^2.447.0",
     "babel-runtime": "6.26.0",
     "base-n": "git://github.com/yourcodesucks/base-n.git",
     "basil-tezos-ledger": "0.0.1",
-    "bip32": "^1.0.0",
-    "bip39": "^2.5.0",
+    "big-integer": "^1.6.43",
+    "bip39": "^3.0.1",
+    "blakejs": "^1.1.0",
     "bs58check": "^2.1.2",
     "https": "^1.0.0",
-    "libsodium-wrappers": "^0.5.4",
-    "libsodium-wrappers-sumo": "^0.5.4",
-    "node-fetch": "^2.3.0",
+    "libsodium-wrappers-sumo": "0.7.4",
+    "moo": "^0.5.0",
+    "nearley": "^2.16.0",
+    "node-fetch": "^2.5.0",
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
     "@types/chai": "^4.1.3",
-    "@types/mocha": "^5.2.0",
+    "@types/chai-as-promised": "^7.1.0",
+    "@types/mocha": "^5.2.6",
     "@types/nock": "^9.3.1",
-    "@types/node": "^10.12.19",
-    "@types/node-fetch": "^2.1.4",
+    "@types/node": "^11.13.8",
+    "@types/node-fetch": "^2.3.3",
     "@types/zxcvbn": "^4.4.0",
+    "@typescript-eslint/eslint-plugin": "^1.7.0",
+    "@typescript-eslint/parser": "^1.7.0",
     "awesome-typescript-loader": "^5.0.0",
     "chai": "^4.1.2",
-    "coveralls": "^3.0.2",
-    "eslint": "^5.12.0",
+    "chai-as-promised": "^7.1.1",
+    "copyfiles": "^2.1.0",
+    "coveralls": "^3.0.3",
+    "eslint": "^5.16.0",
     "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-plugin-import": "^2.16.0",
+    "eslint-plugin-import": "^2.17.2",
     "eslint-plugin-typescript": "^0.14.0",
-    "mocha": "^5.2.0",
+    "mocha": "^6.1.4",
     "nock": "^10.0.6",
-    "nyc": "^13.1.0",
+    "nyc": "^14.0.0",
+    "request-promise": "^4.2.4",
     "source-map-loader": "^0.2.4",
-    "source-map-support": "^0.5.10",
-    "ts-loader": "^4.3.1",
-    "ts-node": "^6.0.5",
+    "source-map-support": "^0.5.12",
+    "terser-webpack-plugin": "^1.2.3",
+    "ts-loader": "^5.4.5",
+    "ts-node": "^8.1.0",
     "tsconfig-paths-webpack-plugin": "^3.2.0",
-    "typedoc": "^0.11.1",
-    "typescript": "^2.8.3",
-    "typescript-eslint-parser": "^21.0.2",
-    "uglifyjs-webpack-plugin": "^1.2.5",
-    "webpack": "^4.29.0",
-    "webpack-cli": "^3.2.1",
+    "typedoc": "^0.14.2",
+    "typedoc-plugin-markdown": "^1.2.0",
+    "typescript": "^3.4.5",
+    "webpack": "^4.30.0",
+    "webpack-cli": "^3.3.1",
     "webpack-node-externals": "^1.7.2"
   },
-  "peerDependencies": {
-    "node-fetch": "^2.3.0"
-  },
   "eslintConfig": {
-    "parser": "typescript-eslint-parser",
+    "parser": "@typescript-eslint/parser",
+    "plugins": [
+      "@typescript-eslint"
+    ],
     "env": {
       "mocha": true,
       "node": true
