@@ -1,34 +1,41 @@
-const UglifyJsPlugin = require('uglifyjs-webpack-plugin')
-const path = require('path')
-const TsConfigPathsPlugin = require('tsconfig-paths-webpack-plugin')
-const nodeExternals = require('webpack-node-externals');
+const TerserPlugin = require('terser-webpack-plugin');
+const path = require('path');
+const TsConfigPathsPlugin = require('tsconfig-paths-webpack-plugin');
+const { CheckerPlugin } = require('awesome-typescript-loader');
 
-module.exports = {
-    mode: "development",
-    devtool: "inline-source-map",
-    entry: "./src/index.ts",
-    target: 'node',
+const webConfig = {
+    mode: 'production',
+    entry: './src/index-web.ts',
+    target: 'web',
     output: {
-        path: path.resolve(__dirname, './build'),
-        filename: 'bundle.js',
+        path: path.resolve(__dirname, './dist-web'),
+        filename: 'conseiljs.min.js',
         library: 'conseiljs',
         libraryTarget: 'umd'
     },
     resolve: {
-        extensions: [".ts", ".tsx", ".js"],
+        extensions: ['.ts', '.tsx', '.js'],
         plugins: [
             new TsConfigPathsPlugin({
-                configFile: './tsconfig.json',
-            }),
-        ],
+                configFile: './tsconfig.json'
+            })
+        ]
     },
     module: {
         rules: [
             { test: /\.tsx?$/, loader: 'awesome-typescript-loader' }
         ]
     },
-    externals: [nodeExternals()],
-    plugins: [
-        new UglifyJsPlugin()
-    ]
+    node: {
+        child_process: 'empty',
+        fs: 'empty',
+        net: 'empty',
+        tls: 'empty'
+    },
+    plugins: [new CheckerPlugin()],
+    optimization: {
+        minimizer: [new TerserPlugin()]
+    }
 };
+
+module.exports = [webConfig];
